+++ tests/testthat/test-arg.R
@@ -1,7 +1,7 @@
 context("arg")
 
 test_that("deprecated() returns the missing argument", {
-  fn <- function(foo = deprecated()) is_supplied(foo)
+  fn <- function(foo = deprecated()) is_present(foo)
   expect_false(fn())
   expect_true(fn(1))
 })